Move ClosureUtils into ClosureCompiler
