+++ spaghetti-core/src/test/groovy/com/prezi/spaghetti/obfuscation/ClosureCompilerTest.groovy
@@ -175,4 +175,41 @@ console.log(Entry.EntryValue);
         then:
         ret == 1
     }
+
+    def "test single entry point"() {
+        when:
+        File dir = Files.createTempDirectory("ClosureUtilsTest").toFile();
+        dir.mkdirs();
+
+        def mainFile = new File(dir, "main.js");
+        ClosureCompiler.writeMainEntryPoint(
+            mainFile,
+            [ new File(dir, "folder/folder/module.js") ],
+            "prezi_closure_test");
+
+        then:
+        mainFile.text == "prezi_closure_test=require('./folder/folder/module');"
+    }
+
+    def "test multiple entry points"() {
+        when:
+        File dir = Files.createTempDirectory("ClosureUtilsTest").toFile();
+        dir.mkdirs();
+
+        def mainFile = new File(dir, "main.js");
+        ClosureCompiler.writeMainEntryPoint(
+            mainFile,
+            [
+                new File(dir, "folder/folder/module.js"),
+                new File(dir, "module.js"),
+                new File(dir.getParentFile(), "module.js"),
+            ],
+            "prezi_closure_test");
+
+        then:
+        mainFile.text == ("prezi_closure_test=[" +
+            "require('./folder/folder/module')," +
+            "require('./module')," +
+            "require('../module')];")
+    }
 }
